[ci] use GitHub app token to create pull request
This modifies the cherry-pick workflow to use a lowRISC-managed GitHub
token minter to generate an access token with ability to create
pull_requests as the "lowrisc-ci" app.

This allows GitHub action to trigger and bypass the restriction that
actions performed by GitHub actions cannot itself trigger another GitHub
actions (as a recursion prevention measure), and therefore allows CI to
run on cherry-picked pull requests.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>