[csrng/rtl] Remove the cmdreq FIFO from ctr_drbg_cmd
Remove the cmdreq FIFO in csrng_ctr_drbg_cmd and instead use data from
the state db and the adata unpacker. The FIFO in question is the largest
in the design, measuring around 9kGE.

This commit can be seen as a blueprint for the ones that follow in
regards to the removal of the respective error reporting circuitry and
bits in the register map, dv, and dif.

Signed-off-by: Florian Glaser <glaserf@lowrisc.org>